Initial cyborg api and db design proposal
This spec proposes the initial design for the
cyborg api. The cyborg api should support the basic
operations concerning accelerators, and does not
necessarily have to be user facing api at the early
stage.The api should support functionalities such as
provision, attach, detach, list and update.

This spec also contains the proposal for a simple DB for
Cyborg. Note that although we propose a DB schema for Cyborg,
in implementation it should be aligned with resource provider
db schema as much as possible.

APIImpact

Change-Id: I98c74df91f4548ecef42d2e3f96facf9023a346a
Signed-off-by: zhipengh <huangzhipeng@huawei.com>