Cyborg Agent Spec
Cyborg Agent resides on various compute hosts and monitors them for accelerators.
If an accelerator is present but not setup Cyborg Agent will notify the operator
that intervention is required. If the accelerators are prepared for use Cyborg
Agent will begin monitoring their status and reporting it to Cyborg Conductor
which will manage use this information to assist with scheduling and operation.

Change-Id: I9596297b8fc64571035d800a24c3e54142dfb0e2