Agent Stubs
So I think this has the most functionality we can expect of a stub
at this point, it accepts a configured transport URL and node id
then takes rpc commands to modify the state of the agent.

What needs to be added is a accelerator handler class that runs
in the main loop to perform the various ongoing tasks of the Agent
but I can't really write that until I have a driver to link against.

Change-Id: I35438c01ca2765ca3dd2575bf2fac63099a30d29