Database interface using sqlalchemy
This commit adds sqlalchemy as a database backend, the database connection
url is loaded from the config file in the same format as any other service.

The accelerator class has been given tags that allow sqlalchemy to map it to
a table in the database and perform saving out/queries automatically.

For further work I'll want either an api or a driver to start integrating with
don't want to get too deep into buisness logic before we know exactly what those
will look like.

Change-Id: I6a9712de0b4f985bab80178845f59e10c05032a5