Modify tox.ini to reflect patch tests
So right now flake8 is in our tox.ini but not in the upstream
infra config. Meaning if you run tox locally with the hope of
finding and fixing linter errors in your patch you get about
100 lines of docstring errors that aren't actually enforced
when you upload your patch to gerrit. We need to either chose
to enforce those rules on gerrit (and write a lot of docstrings)
or change the tox.ini to reflect the actual gates better so that
we can more easily debug locally.

Change-Id: Id0456f53d176eb1d13c3bc22ef43b62772a86b89