Cyborg Nova interaction take 2
Dramatically revised from this

https://review.openstack.org/#/c/448228/3/specs/proposal/cyborg-nova-interaction.rst

My goal with this spec is to have it be only where we interact with nova without
details of other components, otherwise we just get into a monolithic spec for
everything.

I plan to expand this into exact api calls and a detailed workflow, especially
for the new API call we will have to make in nova to register whitelisted
devices live.  That being said we may need to reboot the machine to change the
grub config anyways so maybe we should be looking at how to make that work
first.

Change-Id: I22037109b613d7b33d7c620b78493ec7e96e735e