Fix pep8 error in cyborg/*.py and add Forbidden class
This patch will first fix pep8 error in cyborg/*.py, but except
the sub dir, like cyborg/api*、cyborg/cmd/*...., these code will
be fixed in the follow patch.

This patch does two things:

1. Add Forbidden class in cyborg/common/excpetion.py and modify
NotAuthorized class、HTTPForbidden class.

For the cyborg/context.py, There is no exception called Forbidden
in the cyborg library. I guess this part of the code is referenced
to nova[1]. However, in the nova library [2], they define a Forbidden
class, so we can also defined class Forbidden in cyborg/common/exception.py,
and import it in the format of pep8.

At the same time, the content defined by NotAuthorized and
HTTPForbidden is inconsistent with the effect it wants to
achieve, so NotAuthorized and HTTPForbidden are modified here.

2. Adjust the import order in cyborg/quota.py.

[1]. https://github.com/openstack/nova/blob/master/nova/context.py#L300
[2]. https://github.com/openstack/nova/blob/master/nova/exception.py#L162

Change-Id: Ia7b22272412b5e5fe6a5fa0cdf06fff7870bc07d