[train][goal] Run 'cyborg-tempest-ipv6-only' job in gate
As part of Train community goal 'Support IPv6-Only Deployments and Testing'[1],
Tempest has defined the base job 'devstack-tempest-ipv6' which will
deploy services on IPv6.

This commit Run 'cyborg-tempest-ipv6-only' job in cyborg gate.

Because cyborg-tempest-ipv6-only is non-voting job it is added on check
pipeline only. This is non voting because cyborg-tempest job is not yet
voting.

Depends-On: https://review.opendev.org/#/c/682514/

Story: #2005477
Task: #35886

[1] https://governance.openstack.org/tc/goals/train/ipv6-support-and-testing.html

Change-Id: I24e6e1e407391932049a126502ba94e57cbc6124