Update driver section.
Apart from FPGA and GPU, we support other device types
today and will add more in the future.

Most of the operations are common to the drivers we support.
For example, all drivers support device discovery, VM
creation/deletion, etc. None of them support firmware/shell
update today.

Also, we need to include the supported product families and
any special notes for the driver.

So this update removes the per-device-type files for FPGA
and GPU, and instead adds per-driver notes.

Change-Id: Ic0836e5a8d470a0f185d8d73e4f9ae21dfa63e45