bug fix: ensure resource class instead of getting it.
Before, we used "get" to make sure the resource class exist in our
enviroment. If we customerize a new resource class like
"CUSTOM_AICHIP", the "get" function will return a 404 without throwing
out an exception, the ensure_resource_class will not be called. So we
will fail to create a new resource class.

This patch fixes this bug.

Change-Id: I203fe44ab9996c6c1e0e5ed85655674ed1119824