Bugfix: _needs_programming lack one situation process
The situation:
First, we create one vm with bitstream successfully and then we delete it.

Then we updated the device_profile without bitstream, re-created the vm
and will report an error. Please see the Story link for specific errors.

The reason for the error is that there is a lack of judge, resulting in
method _needs_programming return True which should return False.

Story: 2006767

Change-Id: I7ac8f164b748e9f54d8fd82737b7d8d37cbd8030