BugFix: Modify objects/device.py get_by_device_id method
This get_by_device_id will returns [], regardless of whether
a device exists in the database. Because there is no device_id
field in the device table.

This method has not yet been executed, so this error is hidden.

Because the id of device is unique, we can use get instead of
list.

Change-Id: I6410d7201c938b67cd5a651447a7d8e661e80a15
Story: 2007038
Task: 37846