Update gpu driver
Current gpu driver is out of data, it cannot generate attribute info
("resource_class"/"rc" and "traits") of the gpu device, as a result
cyborg-conductor cannot report the device to placement and nova cannot
schedule this device during the process of nova boot.
This patch added the attribute_list generation in gpu driver so that
the device info can be reported to placement and user can boot an instance
with gpu device.

Change-Id: Ifd44d388d8982f90a3e1c5dd35116cae68e80627