Send a separate bind event to Nova for each ARQ in an instance.
This is based on discussion with Nova community. See:
https://review.opendev.org/#/c/692707/6/nova/objects/external_event.py@36

Each event has a unique tag, i.e. the ARQ UUID, and the
bind status for that ARQ.

Each ARQ has its own state. However, the bind status sent to Nova
should be 'completed' or 'failed'. The logic to do that conversion
should not be in nova_client.py, to keep it free of ARQ state details.
So it has been added in get_arq_bind_status() in ext_arq_job.py.

Change-Id: Iddbf9a77196fc42ac82ad1f6d88a4b0732852463