Add obj_make_compatible()
This adds a method to CyborgObject that allows it to convert itself
to and older version, within a compatibility window. So, if an object
had a revision that added or changed the formatting of an attribute,
the obj_make_compatible() method can fix up a primitive representation
before it is sent to a client expecting the older version.

Partial-Implements: blueprint add-description-field-to-device-profiles

Change-Id: I196629059bc32165f161fe9c071a339d63d71c10