Revert "Solve py37 timeout"
Still not find a good solution.
This reverts commit 08af6012710fc9bbbec0f7936abd701d4eccf7db.

The whole history of timeout can be 3 stage:
1. After we enable py37 there are some random timeout as follow:
https://review.opendev.org/#/c/679406/  On Sep 19 11:30 AM
https://review.opendev.org/#/c/688239/  On Oct 12 3:31 PM
https://review.opendev.org/#/c/688231/  On Oct 28 11:38 PM,
  Oct 29 9:11 AM, Oct 29 12:05 PM
https://review.opendev.org/#/c/685542/  On Nov 14 12:31 AM,
   Nov 14 7:42 AM
https://review.opendev.org/#/c/691872/  On Oct 29 11:55 PM
https://review.opendev.org/#/c/690509/  On Nov 15 4:33 PM

2. After this https://review.opendev.org/#/c/688593/ timeout Disappeared
for a while. This patch is merged on Nov 19 5:05 PM.
This patch set a wrong python path evn.

3. After another patch https://review.opendev.org/#/c/696397， timeout
comes out again.
The intention of this patch is to suppress confusing pep8 message, but
it also fix python path evn unintentionally.

The thread pool lib always loop to check if there are new jobs.
The action of testunit framework in py37 is different with py36. If
there are thread pool run, the testunit will never return.

you can a simple test as follow in this file:
def job1(t1=0):
    print(time.time())
    print("sleep %s second" % t1)
    time.sleep(t1)
    print(time.time())
    return "Hello, world"

class TestExtARQObject(base.DbTestCase):
    def test_foo(self):
        print("Test Foo")

    def test_bar(self):
        print("Test bar")

    def test_apply_patch_fpga_arq_monitor_job(self):
        works = utils.ThreadWorks()
        job = works.spawn(job1, 1)
        return job

Change-Id: I398db324563ecdb6e8fe0abb86fd02c1336b467f