support standard devstack plugin cloning
This change removes the non standard mechanism
for specifying the cyborg repo and branch
via CYBORG_REPO and CYBORG_BRANCH.

in-tree devstack modules use the
<PROJECT NAME>_REPO and <PROJECT NAME>_BRANCH
convention for specifying revisions of project to clone
however devstack plugins should only use that convention
for cloning other git repos.

A devstack plugin should never clone the repo that contains
the plugin itself as doing so will replace the code while it
is executing. This can lead to the code that is running and the
code on disk being different if the plugin line and CYBORG_*
options do not match.

This change updates the devstack_setup docs with an example of
testing an unmerged change.

This change removes some commented out code and unneeded checks in
devstack/lib/*

Change-Id: I473ef276197ad6578f64a708f4b66562a2ff1d49