simplify devstack plugin
This change replaces the nested if used
in plugin.sh with nested case statements for
greater readablity.

This change removes an unnecessary check to determine
if services are enabled. By default all cyborg
service are enabled by the plugin. Since the plugin is
not loaded unless the user specifies the enable_plugin
line in there local.conf we can assume they want the plugin
to run and there for can remove the service check.

This change removes the install_agent function and renames
install_cyborg_in_controller function to
check_cyborg_service_deps to better reflect what the function
does.

Change-Id: Ia5ad1235ae4a5a76bcd498a743ca6be6b3634c5c