Fix setting cyborg agent hostname
This removes the devstack plugin line which overrides the CONF.host in
the cyborg agent with the $LOCAL_HOSTNAME. The latter may be different
than what the nova-compute service generates (as the nova devstack
module does not override the host). Since these need to match, just
remove the override here so that both services use the hostname from
the service framework.

Closes Task: #38811

Change-Id: I393ccbf020cc4f80ce78b1ba4564a833abec02bf