Optimize deployble object module
Because the deployable object and the attribute object
are independent, we should completely remove the attribute
setting at the object level.

When the attribute is removed, deployable can use the
method of the base class like other objects (device, attach_handle).
Therefore, this patch also removes the _from_db_object() method.

Change-Id: Ieebc8242b68cbe1dfd4a3bc9081de2bb5f4075a7
Story: 2007369
Task: 38931