Bugfix: expose placement service unavailable before diff cyborg db
We know that cyborg will periodically synchronize data
with placement, but there is a prerequisite for data
synchronization. That is, when Cyborg periodically scans
and finds that the database has changed, it will update
the placement.

There will be an implicit bug, that is, when the placement
service is abnormal (for example: the installation service
is abnormal, the node where the placement service is down),
cyborg will still update its database. When the placement
service is normal, the cyborg database has been updated.
Although the device inspection will be performed in the next
cycle, because it has been reported to the cyborg database before,
there will be no diff at this time, and the placement side will
not be updated. Therefore, this situation will cause the data
stored in the placement side to be dirty data.

The solution is to expose the exception of the placement service,
rather than a simple error prompt. Stop the data diff by throwing
an unreachable service exception. In the next cycle, maybe the
placement service is back to normal, and at this point diff can
proceed smoothly.
The reason why this is done is because the entry for resource
scheduling is in placement. When the placement data is incorrect,
the data on cyborg side will lose its meaning.

Story: 2007393
Task: 38989

Change-Id: Ia9e7a21f36b7154809d5bcf7812bb7fa73eff6df