Deployable V2 API implementation
This patch aims at implementing basic functions of v2/deployable API.
It contains:
1. Implement deployable list/show API based on v1 deployable API.
2. Remove 'patch' method, because Cyborg V2 API provides other
API resource(ARQ) to let caller to do binding, not based on deployable.
3. Remove 'delete' method, because deployable can not actually be
deleted, it is reported by each driver on the node.
4. Add related UT.

As for further implementation like program API, it will be done in
another patch.

Change-Id: I4fd12a7d34594a63bd80bdc188ddc977d0a094e7
Task: 2007427