revert device and deployable when resource provider create fail
Destroy device and deployable if resource_provider create failed
when cyborg conductor report data to placement. Otherwise, the
data in cyborg will be disaccord with placement.
Just revert driver device and deployable created in Cyborg and rp created
in Placement to reduce the risk of data inconsistency now
between Cyborg and Placement, we will consider to fix in V release [1].

[1] https://etherpad.opendev.org/p/cyborg-victoria-goals the 4th ponit.

Add unit test for destory device and deployable.

Change-Id: I6467981ed5a2196d50d824f92248205d37bb0b92
story: 2007534