Change bandit job from non-voting to voting
Bandit check is a very important tool to check code security.
Before this patch, we put it as a non-voting job to temporarily ignore
existed bandit failures. After fixed all the existed bandit failures,
we should change this test job from non-voting to voting to make the
bandit job actually work, that is to let zuul do the bandit check for
all patches and give them zuul -1  once the new changes fail to pass
bandit the check.

Depends-On: https://review.opendev.org/#/c/720475
Depends-On: https://review.opendev.org/#/c/720456
Depends-On: https://review.opendev.org/#/c/720149
Depends-On: https://review.opendev.org/#/c/720143

Change-Id: I8480243382f66779a12b1e433b69d9be512e843e