Use uwsgi binary from path and docs job fix
All these uwsgi invocations assume that the uwsgi binary is in the
same directory as their project binaries are installed into (probably
/usr/bin).  That may not be correct -- for example if using a packaged
uwsgi on Fedora the binary will live in /usr/sbin/uwsgi (not /usr/bin
where the project files from pip are).

Switch invocations to just find it in the path.
Also fix the docs job

Change-Id: Ib56b4d3f4962466ea9e937b314e54014070e7adc