Optimize microversion support
This changes is prepare to use microversion. This way is better to use
version number and record the microversion change history in this architecture.
If we want to add a microversion, we just need to add a constant parameter
in [1](e.g. MINOR_1_PROJECT_ID=1) and update the MINOR_MAX_VERSION to
MINOR_1_PROJECT_ID.

[1] https://review.opendev.org/#/c/746282/6/cyborg/api/controllers/v2/versions.py@27

Change-Id: If2a2bc8178a6ee073f9e7cc6ac65df51f3034fa2