keep GPU driver's deployable_name unique
The deployable_name should be unique because deployable_name will be
reported to placement as the resource_provider_name, which is required
by placement to be unique. The current naming method for deployable_name
in cyborg cannot ensure the uniqueness, so we should enhance this.

This patch enhances the GPU driver naming method. Others will come
in series.

Change-Id: Ic97671d11e9bcb3648d5385df2c1075b1d833435
Story: 2008074
Task: 40762