keep FPGA driver's deployable_name unique
The deployable_name should be unique because deployable_name will be
reported to placement as the resource_provider_name, which is required
by placement to be unique. The current naming method for deployable_name
in cyborg cannot ensure the uniqueness, so we should enhance this.

This patch enhances the FPGA driver naming method.

Change-Id: If8afa4718c0c90ac938ec691eff699aaf0e5125e
Story: 2008074
Task: 40763