Add trait check for POST Device Profile
At present, when creating a device profile, the function
_validate_post_request only checks if the keys of groups
matches that in ["resources:", "trait:", "accel:"] without
checking the string following "trait:","resources:",this
introduces risks.

For example, the string followed by "trait:" default should be
"CUSTOM_", what follows "resources:" should be a valid resource type,
but the fact is that users may make typos such as adding extra spaces
like "resources:  FPGA", or other format errors. This kind of error
cannot be detected by the current _validate_post_request, and it won't
be exposed until nova parses the rc of cyborg.

This is not in compliance with the process specification and potentially
introduces greater risks. Therefore, it is necessary to implement a more
strict inspection of rc in device profile POST API.

This patch added the following,rc check will be followed in the next
patch.
1.Check the legality of traits name
2.Space removal in traits
3.Trait value check

Change-Id: Ie3a2c6340e2b41edd8d943ac9e0bda0a1e504d3c
Story: 2008143
Task: 40884