Add intel NIC driver
This patch implements a new driver for Intel Nic Card. It
can discover the device and report it to Placement service with
CUSTOM_NIC resource class and specific traits.

Operator should specify the device and correspond profile in a
config file so that nic's driver can read the info from this
file.

Please check the test report in the following link:
https://wiki.openstack.org/wiki/Cyborg/TestReport/IntelNic

Change-Id: Ida0ba8f24b9e226da7f3d7a85fc372247e5281a5
Implements: blueprint sriov-smartnic-support