tools: Allow check-cherry-picks.sh to be disabled by an env var
The checks performed by this script aren't always useful to downstream
consumers of the repo so allow them to disable the script without having
to make changes to tox.ini.

We use the $commit_hash variable for the check for stable-only patches,
else it will incorrectly fail because it is checking the merge patch's
commit message.

Change-Id: I553fdda626b8afb8949d5ae9fc170caf87d31032