Revert device_profile key to show device profile API
This patch[1] remove the 'device_profile' key in the get device profile
api, which affects the stable/* branch's CI job such as [2], and the
'device_profile' can not be removed from the response. In the future,
we can use add a microversion to support removing it, that can be
unified with other APIs like devices and arqs.

[1] https://review.opendev.org/c/openstack/cyborg/+/751988
[2] https://review.opendev.org/c/openstack/cyborg/+/782134

Depends-On: https://review.opendev.org/c/openstack/cyborg-tempest-plugin/+/784527

Closes-Bug: #1922183
Change-Id: I03ccdd6799f25e5ffa987d36955759705e0750b3
(cherry picked from commit 008c3668c04d3f98362fed9502ffaba731d15763)