imporve api "delete arq by uuids"
Cyborg provides an API to delete arqs by passing an arq uuid list.
But this API will exit with an error if the arq uuid is not existing
and will ignore the following arq uuis with no deletion operation.

This patch allow user to delete all existing arqs and return a error
at the end if there is no-existing arq in the input list, without
interrupt the whole operation.

Change-Id: I98f56ab0cf3b5e7661226c3fe50cb75d8669b4ae