Add support for pyproject.toml and wsgi module paths
pip 23.1 removed the "setup.py install" fallback for projects that do
not have pyproject.toml and now uses a pyproject.toml which is vendored
in pip [1][2]. pip 24.2 has now deprecated a similar fallback to
"setup.py develop" and plans to fully remove this in pip 25.0 [3][4][5].
pbr supports editable installs since 6.0.0

pip 25.1 has now been released and the removal is complete.
by adding our own minimal pyproject.toml to ensure we are using the
correct build system.

This change also requires that we adapt how we generate our wsgi
entry point. when pyproject.toml is used the wsgi console script is
not generated in an editbale install such as is used in devstck

To adress this we need to refactor our usage of our wsgi applciation
to use a module path instead. This change does not remove
the declaration of our wsgi_script entry point but it shoudl
be considered deprecated and it will be removed in the future.

To unblock the gate the devstack plugin is modifed to to deploy
using the wsgi module instead of the console script.

[1] https://pip.pypa.io/en/stable/news/#v23-1
[2] pypa/pip#8368
[3] https://pip.pypa.io/en/stable/news/#v24-2
[4] pypa/pip#11457
[5] https://ichard26.github.io/blog/2024/08/whats-new-in-pip-24.2/
Closes-Bug: #2109583

Note:
- It switches python_requires to 3.9 in setup.cfg for 2025.1
  to fix the job.
- pin setuptools to the last version to support pkg_resources to fix tox
  job

Depends-On: https://review.opendev.org/c/openstack/requirements/+/976903

Change-Id: I31fb6bff75ce27aedd94f7f73b7bfdd706127fc2
(cherry picked from commit 3724b467b471f9bd97d4b5275ec7be463ad011db)
Signed-off-by: Chandan Kumar (raukadah) <chkumar@redhat.com>