Use the correct package name for microversion-parse
As of Ib97eea1037fac541c8ebadc526ae9c3bb160f685 the
requirements check job nolonger uses pkg_resources
to normalise the package names, this highlighted that
we are not using the actuall name and were relying on
that normalization for microversion-parse.

This change fixes that by using the correct name.

Change-Id: I25595fe1beeeb216053464ec1d36887896557fa3
Signed-off-by: Sean Mooney <work@seanmooney.info>